+++ mycodo/tests/software_tests/conftest.py
@@ -68,10 +68,10 @@ def create_admin_user(mycodo_db_uri):
     """ mycodo_flask exits if there is no user called admin. So we create one """
 
     with session_scope(mycodo_db_uri) as db_session:
-        if not db_session.query(User).filter_by(user_role=1).count():
+        if not db_session.query(User).filter_by(role=1).count():
             logging.info("--> Creating new 'test' user as an admin")
-            db_session.add(User(user_name='test',
-                                user_role=1))
+            db_session.add(User(name='test',
+                                role=1))
             db_session.commit()
         else:
             logging.warning("--> Dirty User DB: Admin user was already setup in: '{uri}'".format(uri=mycodo_db_uri))